requirements.txt
Flask==2.1.1
gunicorn==20.1.0
tensorflow-cpu==2.8.0
